hw4_final upd

+++ src/services/user.service.ts
@@ -1,25 +1,56 @@
-import { IUser } from "../models/IUser";
+import ApiError from "../errors/api-error";
+import { IUser, TypeUserDto } from "../models/IUser";
 import { userRepository } from "../repositories/user.repository";
 
 class UserService {
   public async getItems(): Promise<IUser[]> {
-    return await userRepository.getItems();
+    const users = await userRepository.getItems();
+    if (!users) {
+      throw new ApiError("User not found", 404);
+    }
+    return users;
   }
 
   public async getById(userId: number): Promise<IUser> {
-    return await userRepository.getById(userId);
+    const user = await userRepository.getById(userId);
+    if (!user) {
+      throw new ApiError("User not found", 404);
+    }
+
+    return user;
   }
 
-  public async create(dto: IUser): Promise<IUser> {
-    return await userRepository.create(dto);
+  public async create(dto: TypeUserDto): Promise<IUser> {
+    if (
+      typeof dto.name !== "string" ||
+      typeof dto.age !== "number" ||
+      typeof dto.email !== "string" ||
+      typeof dto.isActive !== "boolean"
+    ) {
+      throw new ApiError("Incorrect entered data", 400);
+    }
+    const user = await userRepository.create(dto);
+    if (!user) {
+      throw new ApiError("User not found", 404);
+    }
+
+    return user;
   }
 
   public async update(dto: IUser, userId: number): Promise<IUser> {
-    return await userRepository.update(dto, userId);
+    const user = await userRepository.update(dto, userId);
+    if (!user) {
+      throw new ApiError("User not found", 404);
+    }
+    return user;
   }
 
-  public async delete(userId: number): Promise<IUser> {
-    return await userRepository.delete(userId);
+  public async delete(userId: number): Promise<void> {
+    const user = await userRepository.getById(userId);
+    if (!user) {
+      throw new ApiError("User not found", 404);
+    }
+    await userRepository.delete(userId);
   }
 }
